Better implementation of Deque, Queue and Stack. Now this methods does not use the full trait but only the methods required
