+++ sources/GenericCollections/Abstracts/AbstractStack.php
@@ -3,30 +3,29 @@
 use GenericCollections\Interfaces\QueueInterface;
 use GenericCollections\Internal\DataDoubleLinkedList;
 use GenericCollections\Traits\CollectionMethods;
-use GenericCollections\Traits\DequeMethods;
-use GenericCollections\Traits\StackLifoMethods;
+use GenericCollections\Traits\DequeCommonMethods;
 
 abstract class AbstractStack extends DataDoubleLinkedList implements QueueInterface
 {
     use CollectionMethods;
 
-    use DequeMethods {
-        addFirst as private;
-        addLast as private;
-        offerFirst as private;
-        offerLast as private;
-        getFirst as private;
-        getLast as private;
-        peekFirst as private;
-        peekLast as private;
-        removeFirst as private;
-        removeLast as private;
-        pollFirst as private;
-        pollLast as private;
+    use DequeCommonMethods {
+        addLast as private;     // not really needed
+        addFirst as private;    // used in add
+        offerLast as private;   // not really needed
+        offerFirst as offer;
+        getFirst as element;
+        peekFirst as peek;
+        removeFirst as remove;
+        pollFirst as poll;
+    }
+
+    public function add($element)
+    {
+        $this->addFirst($element);
+        return true;
     }
     
-    use StackLifoMethods;
-
     protected function containerInternalName()
     {
         return 'stack';
